requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0

# TLOB library dependencies
einops>=0.6.0
hydra-core>=1.3.0
lightning>=2.0.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
transformers>=4.20.0

# Additional utilities
tqdm>=4.64.0
jupyter>=1.0.0
notebook>=6.4.0 